Rename the file to match the new class name
