+++ nuclearmasses/ame_mass_file.py
@@ -1,4 +1,4 @@
-from nuclearmasses.parse import ElementConverter
+from nuclearmasses.element_converter import ElementConverter
 
 
 class AMEMassFile(ElementConverter):
