+++ nuclearmasses/ame_reaction_2_file.py
@@ -1,4 +1,4 @@
-from nuclearmasses.parse import ElementConverter
+from nuclearmasses.element_converter import ElementConverter
 
 
 class AMEReactionFileTwo(ElementConverter):
